+++ src/operations/transactions/ux/vote_ux.c
@@ -30,7 +30,7 @@
 #include "utils/print.h"
 #include "utils/utils.h"
 
-#include "ux/display_context.h"
+#include "display/context.h"
 
 ////////////////////////////////////////////////////////////////////////////////
 
@@ -54,14 +54,14 @@ void displayVote(const Transaction *transaction) {
     bytecpy((char *)displayCtx.title[1], LABEL_FEE, LABEL_FEE_SIZE);
 
     // Vote
-    displayCtx.var[0][0] = voteSymbol;
-    bytesToHex((char *)&displayCtx.var[0][1],
+    displayCtx.text[0][0] = voteSymbol;
+    bytesToHex((char *)&displayCtx.text[0][1],
                 &transaction->asset.vote.data[1],
                 PUBLICKEY_COMPRESSED_LEN);
 
     // Fee
     printAmount(transaction->fee,
-                (uint8_t *)displayCtx.var[1], sizeof(displayCtx.var[1]),
+                (uint8_t *)displayCtx.text[1], sizeof(displayCtx.text[1]),
                 TOKEN_NAME, TOKEN_NAME_SIZE, TOKEN_DECIMALS);
 }